new update + bug fixes
+ upgraded Gradle v5.5 RC-3 to 5.6.1 (Android Studio 3.6 Beta 1)
+ migrated to AndroidX Jetpack
+ added expandable definition view in viewholder
+ removed unused / unnecessary resources
+ removed redundant null-checks and variables
+ removed RecyclerView animation changes (because of crashes)
+ fixed some RecyclerView viewholder bugs
+ fixed Settings dialog option alignment
+ fixed some font and intent search issues
+ some bug fixes and performance improvements

+++ app/src/main/java/awais/backworddictionary/custom/SearchHistoryTable.java
@@ -6,6 +6,8 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
+import androidx.annotation.NonNull;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -13,7 +15,7 @@ import awais.backworddictionary.R;
 import awais.lapism.SearchItem;
 
 public class SearchHistoryTable {
-    private static int mHistorySize = 4;
+    private final static int HISTORY_SIZE = 8;
     private final SearchHistoryDatabase dbHelper;
     private SQLiteDatabase db;
 
@@ -23,79 +25,72 @@ public class SearchHistoryTable {
     }
 
     public void open() {
-        if (db == null && dbHelper != null) db = dbHelper.getWritableDatabase();
-        if (db != null && !db.isOpen() && dbHelper != null) db = dbHelper.getWritableDatabase();
+        if ((db == null || !db.isOpen()) && dbHelper != null) db = dbHelper.getWritableDatabase();
     }
 
     public void close() {
         if (db != null && db.isOpen()) dbHelper.close();
     }
 
-    public void setHistorySize(int historySize) {
-        mHistorySize = historySize;
-    }
-
     public void addItem(SearchItem item) {
-        if (db == null) return;
-        if (!db.isOpen()) return;
-        ContentValues values = new ContentValues();
-        boolean isTextAvailable = checkText(item.get_text().toString());
+        if (db == null || !db.isOpen()) return;
+
+        final ContentValues values = new ContentValues();
+        final boolean isTextAvailable = checkText(item.get_text().toString());
+
         if (!isTextAvailable) {
             values.put("_text", item.get_text().toString());
             db.insert("search_history", null, values);
         } else {
-            int lastItemId = getLastItemId();
+            final int lastItemId = getLastItemId();
             values.put("_id", lastItemId + 1);
-            db.update("search_history", values, "_id = ?", new String[]{Integer.toString(getItemId(item))});
+            db.update("search_history", values, "_id = ?", new String[] {Integer.toString(getItemId(item))});
         }
     }
 
     public List<SearchItem> getAllItems(String databaseKey) {
-        List<SearchItem> list = new ArrayList<>();
-        String selectQuery = "SELECT * FROM search_history";
-        if (databaseKey != null) selectQuery += " WHERE _text = " + databaseKey;
-        selectQuery += " ORDER BY _id DESC LIMIT " + mHistorySize;
+        final List<SearchItem> list = new ArrayList<>();
+        if (db == null || !db.isOpen()) return list;
 
-        if (db == null) return list;
-        if (!db.isOpen()) return list;
+        final Cursor cursor = db.rawQuery(databaseKey == null ?
+                "SELECT * FROM search_history ORDER BY _id DESC LIMIT " + HISTORY_SIZE :
+                "SELECT * FROM search_history WHERE _text = " + databaseKey, null);
 
-        Cursor cursor = db.rawQuery(selectQuery, null);
-        if (cursor.moveToFirst()) do
-            list.add(new SearchItem(R.drawable.ic_history_black_24dp, cursor.getString(1)));
-        while (cursor.moveToNext());
+        if (cursor.moveToFirst())
+            do
+                list.add(new SearchItem(R.drawable.ic_history_black_24dp, cursor.getString(1)));
+            while (cursor.moveToNext());
         cursor.close();
 
         return list;
     }
 
     public void clearDatabase(String key) {
-        if (db == null) return;
-        if (!db.isOpen()) return;
+        if (db == null || !db.isOpen()) return;
+
         if (key == null) db.delete("search_history", null, null);
-        else db.delete("search_history", "_text = ?", new String[]{key});
+        else db.delete("search_history", "_text = ?", new String[] {key});
     }
 
-    @SuppressWarnings("unused")
     public int getItemsCount() {
-        if (db == null) return -1;
-        if (!db.isOpen()) return -1;
-        Cursor cursor = db.rawQuery("SELECT * FROM search_history;", null);
-        int count = cursor.getCount();
+        if (db == null || !db.isOpen()) return -1;
+
+        final Cursor cursor = db.rawQuery("SELECT * FROM search_history;", null);
+        final int count = cursor.getCount();
         cursor.close();
         return count;
     }
 
-    private int getItemId(SearchItem item) {
-        Cursor res = db.rawQuery("SELECT _id FROM search_history WHERE _text = ?;", new String[]{item.get_text().toString()});
+    private int getItemId(@NonNull SearchItem item) {
+        final Cursor res = db.rawQuery("SELECT _id FROM search_history WHERE _text = ?;", new String[] {item.get_text().toString()});
         res.moveToFirst();
-        int id = res.getInt(0);
+        final int id = res.getInt(0);
         res.close();
         return id;
     }
 
     private int getLastItemId() {
-        String sql = "SELECT _id FROM search_history";
-        Cursor res = db.rawQuery(sql, null);
+        final Cursor res = db.rawQuery("SELECT _id FROM search_history", null);
         int count = 0;
         if (res.moveToLast()) count = res.getInt(0);
         res.close();
@@ -103,37 +98,34 @@ public class SearchHistoryTable {
     }
 
     private boolean checkText(String text) {
-        Cursor cursor = db.rawQuery("SELECT _text FROM search_history WHERE _text = ?;", new String[]{text});
-        boolean ret = cursor.moveToFirst();
+        final Cursor cursor = db.rawQuery("SELECT _text FROM search_history WHERE _text = ?;", new String[] {text});
+        final boolean ret = cursor.moveToFirst();
         cursor.close();
         return ret;
     }
 
     private static class SearchHistoryDatabase extends SQLiteOpenHelper {
-//        static final String SEARCH_HISTORY_TABLE = "search_history";
-//        static final String SEARCH_HISTORY_COLUMN_ID = "_id";
-//        static final String SEARCH_HISTORY_COLUMN_TEXT = "_text";
-
         SearchHistoryDatabase(Context context) {
             super(context, "search_history_database.db", null, 4);
         }
 
         @Override
-        public void onCreate(SQLiteDatabase db) {
+        public void onCreate(@NonNull SQLiteDatabase db) {
             db.execSQL("CREATE TABLE IF NOT EXISTS search_history ( _id INTEGER PRIMARY KEY AUTOINCREMENT, _text TEXT );");
         }
 
         @Override
-        public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+        public void onUpgrade(@NonNull SQLiteDatabase db, int oldVersion, int newVersion) {
             db.execSQL("DROP TABLE IF EXISTS search_history");
             onCreate(db);
         }
 
         @Override
-        public void onDowngrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+        public void onDowngrade(@NonNull SQLiteDatabase db, int oldVersion, int newVersion) {
             onUpgrade(db, oldVersion, newVersion);
         }
 
-        @Override public void onOpen(SQLiteDatabase db) {}
+        @Override
+        public void onOpen(SQLiteDatabase db) {}
     }
 }
